Repo Reorganisation and Update
Our work was split across the repo in seperate folders. This push
contains the work Cedric and I have done over the last few days to
collaborate our code with the html provided by the first years. It also
reorganises our existing files along with the added ones to match the
final product hosted on the web server currently.